add sample

+++ ThinningCoordinator/Sources/TCHelper.h
@@ -12,5 +12,5 @@
 
 BOOL TCCollectionViewSupportsConstraintsProperty();
 BOOL TCCollectionViewSupportsConstraintsProperty();
-void TCValidateArrayArgument(NSArray *arr, const int index, const void *file, const int line, const void *method);
+void TCValidateArrayArgument(NSArray *arr, const NSInteger index, const void *file, const int line, const void *method);
 @end
